image/requirements.txt
matplotlib==3.1.0
numpy==1.16.4
PeakUtils==1.3.2
scikit-learn==0.21.2
scipy==1.4.1
sklearn==0.0
speechpy==2.4
colorama==0.3.9
urllib3[secure]==1.24.2
prettytable==0.7.2
junit-xml==1.8
pyyaml==4.2b1
jsonschema==2.6.0
future==0.16.0
six==1.12.0
tensorflow==2.2.0
Pillow==7.1.2
umap-learn==0.4.4
tbb==2019.0
